requirements.txt
Django>=6.0
django-allauth>=65.0
PyJWT>=2.0
requests>=2.28
python-dotenv>=1.0
Pillow>=10.0
gunicorn>=21.0
dj-database-url>=2.0
